requirements.txt
python-keystoneclient
yaml
